@@ -1,4 +1,4 @@
-const _lorem_words = [
+const lorem_words = [
   '活动', '重要', '显示', '大小', '使用', '最后', '系列', '注意', '一些', '其中',
   '我的', '怎么', '最新', '只要', '为了', '一下', '位置', '组织', '日期', '成功',
   '男人', '那些', '需要', '对于', '城市', '成为', '电影', '简介', '免费', '软件',
@@ -35,7 +35,7 @@ const _lorem_words = [
 ];
 
 function word() {
-  return this.pick(this._lorem_words);
+  return this.random(this.data.lorem_words);
 }
 
 function text(len = 200) {
@@ -47,9 +47,12 @@ function text(len = 200) {
   return text.substr(0, len - 1) + '。';
 }
 
-module.exports = function(pino) {
-  pino._lorem_words = _lorem_words;
-
-  pino.register('word', word);
-  pino.register('text', text);
+export default function(pino) {
+  pino.registers({
+    // data
+    lorem_words,
+    // method
+    word,
+    text,
+  });
 };
